requirements.txt
Django~=3.2
cryptography~=38.0
djangorestframework~=3.14